aiap_xgboost_traffic/requirements.txt
matplotlib==3.1.0
numpy==1.17.3
pandas==0.24.2
scikit-learn==0.21.2
seaborn==0.9.0
xgboost==0.90
mysql-connector-python==8.0.17